refactor: remove .common. and .esm from dist scripts
