+++ test/scoring.test.js
@@ -1,4 +1,4 @@
-import Fuse from '../dist/fuse.esm.mjs'
+import Fuse from '../dist/fuse.mjs'
 
 const defaultList = ['Stove', 'My good friend Steve from college']
 const defaultOptions = {}
